Move utility funcs from calcSrd to util
These functions seem fairly generic and would be a better fit being made
available in the util module, which already had a number of similar
functions.
